Move internals of QgsColorDialogV2 to new QgsCompoundColorWidget widget
This allows reuse of the full options from the color dialog when
required in a widget form.
